@@ -51,12 +51,15 @@
     "karma-webpack": "~1.7.0",
     "mocha": "~2.3.3",
     "phantomjs": "^1.9.18",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^15.5.0",
+    "react-dom": "^15.5.0",
     "webpack": "~1.14.0"
   },
   "peerDependencies": {
-    "babel-runtime": ">=5.8.0"
+    "babel-runtime": ">=5.8.0",
+    "react": ">=15.5.0"
+  },
+  "dependencies": {
+    "prop-types": ">=15.5.0"
   }
 }
